+++ climpred/tests/test_checks.py
@@ -1,6 +1,5 @@
 import numpy as np
 import pytest
-import xarray as xr
 
 from climpred.checks import (
     get_chunksize,
@@ -16,42 +15,6 @@ from climpred.checks import (
 from climpred.constants import VALID_LEAD_UNITS
 from climpred.exceptions import DatasetError, DimensionError, VariableError
 
-
-@pytest.fixture
-def ds1():
-    return xr.Dataset(
-        {'air': (('lon', 'lat'), [[0, 1, 2], [3, 4, 5], [6, 7, 8]])},
-        coords={'lon': [1, 3, 4], 'lat': [5, 6, 7]},
-    )
-
-
-@pytest.fixture
-def ds2():
-    return xr.Dataset(
-        {'air': (('lon', 'lat'), [[0, 1, 2], [3, 4, 5], [6, 7, 8]])},
-        coords={'lon': [1, 3, 6], 'lat': [5, 6, 9]},
-    )
-
-
-@pytest.fixture
-def da1():
-    return xr.DataArray([[0, 1], [3, 4], [6, 7]], dims=('x', 'y'))
-
-
-@pytest.fixture
-def da2():
-    return xr.DataArray([[0, 1], [5, 6], [6, 7]], dims=('x', 'y'))
-
-
-@pytest.fixture
-def da_lead_time():
-    lead = np.arange(5)
-    init = np.arange(5)
-    return xr.DataArray(
-        np.random.rand(len(lead), len(init)),
-        dims=['init', 'lead'],
-        coords=[init, lead],
-    )
 
 
 @is_xarray(0)
@@ -249,31 +212,31 @@ def test_is_in_list_fail():
 
 
 @pytest.mark.parametrize('lead_units', VALID_LEAD_UNITS)
-def test_valid_lead_units(da_lead_time, lead_units):
+def test_valid_lead_units(da_lead, lead_units):
     """Test that lead units check passes with appropriate lead units."""
-    da_lead_time['lead'].attrs['units'] = lead_units
-    assert has_valid_lead_units(da_lead_time)
+    da_lead['lead'].attrs['units'] = lead_units
+    assert has_valid_lead_units(da_lead)
 
 
-def test_valid_lead_units_no_units(da_lead_time):
+def test_valid_lead_units_no_units(da_lead):
     """Test that valid lead units check breaks if there are no units."""
     with pytest.raises(AttributeError):
-        has_valid_lead_units(da_lead_time)
+        has_valid_lead_units(da_lead)
 
 
-def test_valid_lead_units_invalid_units(da_lead_time):
+def test_valid_lead_units_invalid_units(da_lead):
     """Test that valid lead units check breaks if invalid units provided."""
-    da_lead_time['lead'].attrs['units'] = 'dummy'
+    da_lead['lead'].attrs['units'] = 'dummy'
     with pytest.raises(AttributeError):
-        has_valid_lead_units(da_lead_time)
+        has_valid_lead_units(da_lead)
 
 
 @pytest.mark.parametrize('lead_units', VALID_LEAD_UNITS)
-def test_nonplural_lead_units_works(da_lead_time, lead_units):
+def test_nonplural_lead_units_works(da_lead, lead_units):
     """Test that non-plural lead units work on lead units check."""
-    da_lead_time['lead'].attrs['units'] = lead_units[:-1]
+    da_lead['lead'].attrs['units'] = lead_units[:-1]
     with pytest.warns(UserWarning) as record:
-        has_valid_lead_units(da_lead_time)
+        has_valid_lead_units(da_lead)
     expected = f'The letter "s" was appended to the lead units; now {lead_units}.'
     assert record[0].message.args[0] == expected