+++ CHANGELOG.rst
@@ -2,6 +2,7 @@
 What's New
 ==========
 
+
 climpred v2.0.1 (2020-01-##)
 ============================
 
@@ -16,7 +17,14 @@ New Features
 
     *  user gets warned when chunking potentially (un)-necessary
 
+
+Internals/Minor Fixes
+---------------------
+- Gather all ``pytest.fixture``s in ``conftest.py``. (:pr:`313`) `Aaron Spring`_.
+- Move ``x_METRICS`` and ``COMPARISONS`` to ``metrics.py`` and ``comparisons.py`` in
+  order to avoid circular import dependencies. (:pr:`315`) `Aaron Spring`_.
 - ``asv`` benchmarks for ``HindcastEnsemble`` (:pr:`285`) `Aaron Spring`_.
+
 
 climpred v2.0.0 (2020-01-22)
 ============================